refactor(react-quill): moved custom video module for beter structure
